requirements.txt
altgraph==0.17.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cos-python-sdk-v5==1.9.20
crcmod==1.7
dicttoxml==1.7.4
Flask==2.2.2
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.2
MarkupSafe==2.1.1
modulegraph==0.19.3
proxy-tools==0.1.0
psutil==5.9.2
pycryptodome==3.15.0
pyobjc-core==8.5.1
pyobjc-framework-Cocoa==8.5.1
pyobjc-framework-WebKit==8.5.1
pywebview==3.7
requests==2.28.1
six==1.16.0
tomli==2.0.1
tomli_w==1.0.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
py2app; sys_platform == 'darwin'
pyinstaller; sys_platform != 'darwin'
